requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
backports.entry-points-selectable==1.1.0
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
cryptography==2.6.1
cupshelpers==1.0
distlib==0.3.2
distro==1.3.0
distro-info==0.21
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
httpie==0.9.8
httplib2==0.11.3
idna==2.6
importlib-metadata==4.6.1
keyring==17.1.1
keyrings.alt==3.1.1
louis==3.8.0
Mako==1.0.7
MarkupSafe==1.1.0
olefile==0.46
pexpect==4.6.0
Pillow==5.4.1
platformdirs==2.0.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.30.4
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2021.1
pyxdg==0.25
reportbug===7.5.3-deb10u1
reportlab==3.5.13
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
unattended-upgrades==0.1
urllib3==1.24.1
virtualenv==20.6.0
zipp==3.5.0
